Move rest of major types into core.rs
