+++ druid/src/lib.rs
@@ -24,6 +24,7 @@ pub use druid_shell::{kurbo, piet};
 mod app;
 mod app_delegate;
 mod command;
+mod core;
 mod data;
 mod env;
 mod event;
@@ -37,14 +38,6 @@ mod widget_pod;
 mod win_handler;
 mod window;
 
-use log;
-use std::collections::VecDeque;
-use std::ops::{Deref, DerefMut};
-use std::time::Instant;
-
-use kurbo::{Rect, Size};
-use piet::Piet;
-
 // these are the types from shell that we expose; others we only use internally.
 pub use shell::{
     Application, Clipboard, ClipboardFormat, Cursor, FileDialogOptions, FileInfo, FileSpec,
@@ -52,6 +45,7 @@ pub use shell::{
     TimerToken, WinCtx, WindowHandle,
 };
 
+pub use crate::core::{BoxConstraints, EventCtx, LayoutCtx, PaintCtx, UpdateCtx, Widget};
 pub use app::{AppLauncher, WindowDesc};
 pub use app_delegate::{AppDelegate, DelegateCtx};
 pub use command::{sys as commands, Command, Selector};
@@ -62,542 +56,6 @@ pub use lens::{Lens, LensExt, LensWrap};
 pub use localization::LocalizedString;
 pub use menu::{sys as platform_menus, ContextMenu, MenuDesc, MenuItem};
 pub use mouse::MouseEvent;
-pub use widget_pod::{BaseState, WidgetPod};
+pub use widget_pod::{BaseState, BoxedWidget, WidgetPod};
 pub use win_handler::DruidHandler;
 pub use window::{Window, WindowId};
-
-/// Convenience type for dynamic boxed widget.
-pub type BoxedWidget<T> = WidgetPod<T, Box<dyn Widget<T>>>;
-
-/// The trait implemented by all widgets.
-///
-/// All appearance and behavior for a widget is encapsulated in an
-/// object that implements this trait.
-///
-/// The trait is parametrized by a type (`T`) for associated data.
-/// All trait methods are provided with access to this data, and
-/// in the case of `event` the reference is mutable, so that events
-/// can directly update the data.
-///
-/// Whenever the application data changes, the framework traverses
-/// the widget hierarchy with an [`update`] method. The framework
-/// needs to know whether the data has actually changed or not, which
-/// is why `T` has a [`Data`] bound.
-///
-/// All the trait methods are provided with a corresponding context.
-/// The widget can request things and cause actions by calling methods
-/// on that context.
-///
-/// In addition, all trait methods are provided with an environment
-/// ([`Env`](struct.Env.html)).
-///
-/// Container widgets will generally not call `Widget` methods directly
-/// on their child widgets, but rather will own their widget wrapped in
-/// a [`WidgetPod`], and call the corresponding method on that. The
-/// `WidgetPod` contains state and logic for these traversals. On the
-/// other hand, particularly light-weight containers might contain their
-/// child `Widget` directly (when no layout or event flow logic is
-/// needed), and in those cases will call these methods.
-///
-/// As a general pattern, container widgets will call the corresponding
-/// `WidgetPod` method on all their children. The `WidgetPod` applies
-/// logic to determine whether to recurse, as needed.
-///
-/// [`event`]: #tymethod.event
-/// [`update`]: #tymethod.update
-/// [`Data`]: trait.Data.html
-/// [`WidgetPod`]: struct.WidgetPod.html
-pub trait Widget<T> {
-    /// Handle an event.
-    ///
-    /// A number of different events (in the [`Event`] enum) are handled in this
-    /// method call. A widget can handle these events in a number of ways:
-    /// requesting things from the [`EventCtx`], mutating the data, or submitting
-    /// a [`Command`].
-    ///
-    /// [`Event`]: struct.Event.html
-    /// [`EventCtx`]: struct.EventCtx.html
-    /// [`Command`]: struct.Command.html
-    fn event(&mut self, ctx: &mut EventCtx, event: &Event, data: &mut T, env: &Env);
-
-    /// Handle a change of data.
-    ///
-    /// This method is called whenever the data changes. When the appearance of
-    /// the widget depends on data, call [`invalidate`] so that it's scheduled
-    /// for repaint.
-    ///
-    /// The previous value of the data is provided in case the widget wants to
-    /// compute a fine-grained delta. Before any paint operation, this method
-    /// will be called with `None` for `old_data`. Thus, this method can also be
-    /// used to build resources that will be retained for painting.
-    ///
-    /// [`invalidate`]: struct.UpdateCtx.html#method.invalidate
-
-    // Consider a no-op default impl. One reason against is that containers might
-    // inadvertently forget to propagate.
-    fn update(&mut self, ctx: &mut UpdateCtx, old_data: Option<&T>, data: &T, env: &Env);
-
-    /// Compute layout.
-    ///
-    /// A leaf widget should determine its size (subject to the provided
-    /// constraints) and return it.
-    ///
-    /// A container widget will recursively call [`WidgetPod::layout`] on its
-    /// child widgets, providing each of them an appropriate box constraint,
-    /// compute layout, then call [`set_layout_rect`] on each of its children.
-    /// Finally, it should return the size of the container. The container
-    /// can recurse in any order, which can be helpful to, for example, compute
-    /// the size of non-flex widgets first, to determine the amount of space
-    /// available for the flex widgets.
-    ///
-    /// For efficiency, a container should only invoke layout of a child widget
-    /// once, though there is nothing enforcing this.
-    ///
-    /// The layout strategy is strongly inspired by Flutter.
-    ///
-    /// [`WidgetPod::layout`]: struct.WidgetPod.html#method.layout
-    /// [`set_layout_rect`]: struct.LayoutCtx.html#method.set_layout_rect
-    fn layout(&mut self, ctx: &mut LayoutCtx, bc: &BoxConstraints, data: &T, env: &Env) -> Size;
-
-    /// Paint the widget appearance.
-    ///
-    /// The widget calls methods on the `render_ctx` field of the
-    /// `paint_ctx` in order to paint its appearance. `paint_ctx` auto
-    /// derefs to `render_ctx` for convenience.
-    ///
-    /// Container widgets can paint a background before recursing to their
-    /// children, or annotations (for example, scrollbars) by painting
-    /// afterwards. In addition, they can apply masks and transforms on
-    /// the render context, which is especially useful for scrolling.
-    fn paint(&mut self, paint_ctx: &mut PaintCtx, base_state: &BaseState, data: &T, env: &Env);
-}
-
-// TODO: explore getting rid of this (ie be consistent about using
-// `dyn Widget` only).
-impl<T> Widget<T> for Box<dyn Widget<T>> {
-    fn event(&mut self, ctx: &mut EventCtx, event: &Event, data: &mut T, env: &Env) {
-        self.deref_mut().event(ctx, event, data, env)
-    }
-
-    fn update(&mut self, ctx: &mut UpdateCtx, old_data: Option<&T>, data: &T, env: &Env) {
-        self.deref_mut().update(ctx, old_data, data, env);
-    }
-
-    fn layout(&mut self, ctx: &mut LayoutCtx, bc: &BoxConstraints, data: &T, env: &Env) -> Size {
-        self.deref_mut().layout(ctx, bc, data, env)
-    }
-
-    fn paint(&mut self, paint_ctx: &mut PaintCtx, base_state: &BaseState, data: &T, env: &Env) {
-        self.deref_mut().paint(paint_ctx, base_state, data, env);
-    }
-}
-
-/// A context passed to paint methods of widgets.
-///
-/// Widgets paint their appearance by calling methods on the
-/// `render_ctx`, which PaintCtx derefs to for convenience.
-/// This struct is expected to grow, for example to include the
-/// "damage region" indicating that only a subset of the entire
-/// widget hierarchy needs repainting.
-pub struct PaintCtx<'a, 'b: 'a> {
-    /// The render context for actually painting.
-    pub render_ctx: &'a mut Piet<'b>,
-    pub window_id: WindowId,
-    /// The currently visible region.
-    pub(crate) region: Region,
-}
-
-/// A region of a widget, generally used to describe what needs to be drawn.
-#[derive(Debug, Clone)]
-pub struct Region(Rect);
-
-impl Region {
-    /// Returns the smallest `Rect` that encloses the entire region.
-    pub fn to_rect(&self) -> Rect {
-        self.0
-    }
-
-    /// Returns `true` if `self` intersects with `other`.
-    #[inline]
-    pub fn intersects(&self, other: Rect) -> bool {
-        self.0.intersect(other).area() > 0.
-    }
-}
-
-impl From<Rect> for Region {
-    fn from(src: Rect) -> Region {
-        Region(src)
-    }
-}
-
-impl<'a, 'b: 'a> Deref for PaintCtx<'a, 'b> {
-    type Target = Piet<'b>;
-
-    fn deref(&self) -> &Self::Target {
-        self.render_ctx
-    }
-}
-
-impl<'a, 'b: 'a> DerefMut for PaintCtx<'a, 'b> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        self.render_ctx
-    }
-}
-
-impl<'a, 'b: 'a> PaintCtx<'a, 'b> {
-    /// Returns the currently visible [`Region`].
-    ///
-    /// [`Region`]: struct.Region.html
-    #[inline]
-    pub fn region(&self) -> &Region {
-        &self.region
-    }
-
-    /// Creates a temporary `PaintCtx` with a new visible region, and calls
-    /// the provided function with that `PaintCtx`.
-    ///
-    /// This is used by containers to ensure that their children have the correct
-    /// visible region given their layout.
-    pub fn with_child_ctx(&mut self, region: impl Into<Region>, f: impl FnOnce(&mut PaintCtx)) {
-        let PaintCtx {
-            render_ctx,
-            window_id,
-            ..
-        } = self;
-        let mut child_ctx = PaintCtx {
-            render_ctx,
-            window_id: *window_id,
-            region: region.into(),
-        };
-        f(&mut child_ctx)
-    }
-}
-
-/// A context provided to layout handling methods of widgets.
-///
-/// As of now, the main service provided is access to a factory for
-/// creating text layout objects, which are likely to be useful
-/// during widget layout.
-pub struct LayoutCtx<'a, 'b: 'a> {
-    text_factory: &'a mut Text<'b>,
-    window_id: WindowId,
-}
-
-/// A mutable context provided to event handling methods of widgets.
-///
-/// Widgets should call [`invalidate`] whenever an event causes a change
-/// in the widget's appearance, to schedule a repaint.
-///
-/// [`invalidate`]: #method.invalidate
-pub struct EventCtx<'a, 'b> {
-    // Note: there's a bunch of state that's just passed down, might
-    // want to group that into a single struct.
-    win_ctx: &'a mut dyn WinCtx<'b>,
-    cursor: &'a mut Option<Cursor>,
-    /// Commands submitted to be run after this event.
-    command_queue: &'a mut VecDeque<(WindowId, Command)>,
-    window_id: WindowId,
-    // TODO: migrate most usage of `WindowHandle` to `WinCtx` instead.
-    window: &'a WindowHandle,
-    base_state: &'a mut BaseState,
-    had_active: bool,
-    is_handled: bool,
-    is_root: bool,
-}
-
-/// A mutable context provided to data update methods of widgets.
-///
-/// Widgets should call [`invalidate`] whenever a data change causes a change
-/// in the widget's appearance, to schedule a repaint.
-///
-/// [`invalidate`]: #method.invalidate
-pub struct UpdateCtx<'a, 'b: 'a> {
-    text_factory: &'a mut Text<'b>,
-    window: &'a WindowHandle,
-    // Discussion: we probably want to propagate more fine-grained
-    // invalidations, which would mean a structure very much like
-    // `EventCtx` (and possibly using the same structure). But for
-    // now keep it super-simple.
-    needs_inval: bool,
-    window_id: WindowId,
-}
-
-/// Constraints for layout.
-///
-/// The layout strategy for druid is strongly inspired by Flutter,
-/// and this struct is similar to the [Flutter BoxConstraints] class.
-///
-/// At the moment, it represents simply a minimum and maximum size.
-/// A widget's [`layout`] method should choose an appropriate size that
-/// meets these constraints.
-///
-/// Further, a container widget should compute appropriate constraints
-/// for each of its child widgets, and pass those down when recursing.
-///
-/// [`layout`]: trait.Widget.html#tymethod.layout
-/// [Flutter BoxConstraints]: https://api.flutter.dev/flutter/rendering/BoxConstraints-class.html
-#[derive(Clone, Copy, Debug)]
-pub struct BoxConstraints {
-    min: Size,
-    max: Size,
-}
-
-impl BoxConstraints {
-    /// Create a new box constraints object.
-    ///
-    /// Create constraints based on minimum and maximum size.
-    pub fn new(min: Size, max: Size) -> BoxConstraints {
-        BoxConstraints { min, max }
-    }
-
-    /// Create a "tight" box constraints object.
-    ///
-    /// A "tight" constraint can only be satisfied by a single size.
-    pub fn tight(size: Size) -> BoxConstraints {
-        BoxConstraints {
-            min: size,
-            max: size,
-        }
-    }
-
-    /// Create a "loose" version of the constraints.
-    ///
-    /// Make a version with zero minimum size, but the same maximum size.
-    pub fn loosen(&self) -> BoxConstraints {
-        BoxConstraints {
-            min: Size::ZERO,
-            max: self.max,
-        }
-    }
-
-    /// Clamp a given size so that fits within the constraints.
-    pub fn constrain(&self, size: impl Into<Size>) -> Size {
-        size.into().clamp(self.min, self.max)
-    }
-
-    /// Returns the max size of these constraints.
-    pub fn max(&self) -> Size {
-        self.max
-    }
-
-    /// Returns the min size of these constraints.
-    pub fn min(&self) -> Size {
-        self.min
-    }
-
-    /// Whether there is an upper bound on the width.
-    pub fn is_width_bounded(&self) -> bool {
-        self.max.width.is_finite()
-    }
-
-    /// Whether there is an upper bound on the height.
-    pub fn is_height_bounded(&self) -> bool {
-        self.max.height.is_finite()
-    }
-
-    /// Check to see if these constraints are legit.
-    ///
-    /// Logs a warning if BoxConstraints are invalid.
-    pub fn debug_check(&self, name: &str) {
-        if !(0.0 <= self.min.width
-            && self.min.width <= self.max.width
-            && 0.0 <= self.min.height
-            && self.min.height <= self.max.height)
-        {
-            log::warn!("Bad BoxConstraints passed to {}:", name);
-            log::warn!("{:?}", self);
-        }
-    }
-
-    /// Shrink min and max constraints by size
-    pub fn shrink(&self, diff: impl Into<Size>) -> BoxConstraints {
-        let diff = diff.into();
-        let min = Size::new(
-            (self.min().width - diff.width).max(0.),
-            (self.min().height - diff.height).max(0.),
-        );
-        let max = Size::new(
-            (self.max().width - diff.width).max(0.),
-            (self.max().height - diff.height).max(0.),
-        );
-
-        BoxConstraints::new(min, max)
-    }
-}
-
-impl<'a, 'b> EventCtx<'a, 'b> {
-    /// Invalidate.
-    ///
-    /// Right now, it just invalidates the entire window, but we'll want
-    /// finer grained invalidation before long.
-    pub fn invalidate(&mut self) {
-        // Note: for the current functionality, we could shortcut and just
-        // request an invalidate on the window. But when we do fine-grained
-        // invalidation, we'll want to compute the invalidation region, and
-        // that needs to be propagated (with, likely, special handling for
-        // scrolling).
-        self.base_state.needs_inval = true;
-    }
-
-    /// Get an object which can create text layouts.
-    pub fn text(&mut self) -> &mut Text<'b> {
-        self.win_ctx.text_factory()
-    }
-
-    /// Set the cursor icon.
-    ///
-    /// Call this when handling a mouse move event, to set the cursor for the
-    /// widget. A container widget can safely call this method, then recurse
-    /// to its children, as a sequence of calls within an event propagation
-    /// only has the effect of the last one (ie no need to worry about
-    /// flashing).
-    ///
-    /// This method is expected to be called mostly from the [`MouseMoved`]
-    /// event handler, but can also be called in response to other events,
-    /// for example pressing a key to change the behavior of a widget.
-    ///
-    /// [`MouseMoved`]: enum.Event.html#variant.MouseDown
-    pub fn set_cursor(&mut self, cursor: &Cursor) {
-        *self.cursor = Some(cursor.clone());
-    }
-
-    /// Set the "active" state of the widget.
-    ///
-    /// See [`BaseState::is_active`](struct.BaseState.html#method.is_active).
-    pub fn set_active(&mut self, active: bool) {
-        self.base_state.is_active = active;
-        // TODO: plumb mouse grab through to platform (through druid-shell)
-    }
-
-    /// Query the "hot" state of the widget.
-    ///
-    /// See [`BaseState::is_hot`](struct.BaseState.html#method.is_hot).
-    pub fn is_hot(&self) -> bool {
-        self.base_state.is_hot()
-    }
-
-    /// Query the "active" state of the widget.
-    ///
-    /// This is the same state set by [`set_active`](#method.set_active) and
-    /// is provided as a convenience.
-    pub fn is_active(&self) -> bool {
-        self.base_state.is_active
-    }
-
-    /// Returns a reference to the current `WindowHandle`.
-    ///
-    /// Note: we're in the process of migrating towards providing functionality
-    /// provided by the window handle in mutable contexts instead. If you're
-    /// considering a new use of this method, try adding it to `WinCtx` and
-    /// plumbing it through instead.
-    pub fn window(&self) -> &WindowHandle {
-        &self.window
-    }
-
-    /// Set the event as "handled", which stops its propagation to other
-    /// widgets.
-    pub fn set_handled(&mut self) {
-        self.is_handled = true;
-    }
-
-    /// Determine whether the event has been handled by some other widget.
-    pub fn is_handled(&self) -> bool {
-        self.is_handled
-    }
-
-    /// Query the focus state of the widget.
-    ///
-    /// See [`BaseState::has_focus`](struct.BaseState.html#method.has_focus).
-    pub fn has_focus(&self) -> bool {
-        self.base_state.has_focus()
-    }
-
-    /// Request keyboard focus.
-    ///
-    /// Discussion question: is method needed in contexts other than event?
-    pub fn request_focus(&mut self) {
-        self.base_state.request_focus = true;
-    }
-
-    /// Request an animation frame.
-    pub fn request_anim_frame(&mut self) {
-        self.base_state.request_anim = true;
-    }
-
-    /// Request a timer event.
-    ///
-    /// The return value is a token, which can be used to associate the
-    /// request with the event.
-    pub fn request_timer(&mut self, deadline: Instant) -> TimerToken {
-        self.base_state.request_timer = true;
-        self.win_ctx.request_timer(deadline)
-    }
-
-    /// Returns the layout size of the current widget.
-    pub fn size(&self) -> Size {
-        self.base_state.size()
-    }
-
-    /// Submit a [`Command`] to be run after this event is handled.
-    ///
-    /// Commands are run in the order they are submitted; all commands
-    /// submitted during the handling of an event are executed before
-    /// the [`update()`] method is called.
-    ///
-    /// [`Command`]: struct.Command.html
-    /// [`update()`]: trait.Widget.html#tymethod.update
-    pub fn submit_command(
-        &mut self,
-        command: impl Into<Command>,
-        window_id: impl Into<Option<WindowId>>,
-    ) {
-        let window_id = window_id.into().unwrap_or(self.window_id);
-        self.command_queue.push_back((window_id, command.into()))
-    }
-
-    /// Get the window id.
-    pub fn window_id(&self) -> WindowId {
-        self.window_id
-    }
-}
-
-impl<'a, 'b> LayoutCtx<'a, 'b> {
-    /// Get an object which can create text layouts.
-    pub fn text(&mut self) -> &mut Text<'b> {
-        &mut self.text_factory
-    }
-
-    /// Get the window id.
-    pub fn window_id(&self) -> WindowId {
-        self.window_id
-    }
-}
-
-impl<'a, 'b> UpdateCtx<'a, 'b> {
-    /// Invalidate.
-    ///
-    /// See [`EventCtx::invalidate`](struct.EventCtx.html#method.invalidate) for
-    /// more discussion.
-    pub fn invalidate(&mut self) {
-        self.needs_inval = true;
-    }
-
-    /// Get an object which can create text layouts.
-    pub fn text(&mut self) -> &mut Text<'b> {
-        self.text_factory
-    }
-
-    /// Returns a reference to the current `WindowHandle`.
-    ///
-    /// Note: For the most part we're trying to migrate `WindowHandle`
-    /// functionality to `WinCtx`, but the update flow is the exception, as
-    /// it's shared across multiple windows.
-    pub fn window(&self) -> &WindowHandle {
-        &self.window
-    }
-
-    /// Get the window id.
-    pub fn window_id(&self) -> WindowId {
-        self.window_id
-    }
-}
